requirements.txt

fastapi
uvicorn
pillow
pytesseract
python-multipart
camelot-py[cv]
opencv-python-headless
